// src/types.ts
/**
 * If conversation growing very large and want to optimize performance, we can 
 * implementing lazy loading or pagination to load only a subset of the 
 * conversation at a time.
 * Additionally, if we need to persist the conversation data across sessions 
 * or enable features like search or filtering, you can consider storing the 
 * conversation data in a database or a server-side storage solution.
 */
export interface CompleteConverse {
    prompt: string;
    response: {
        type: 'image' | 'text' | 'imageClassification';
        content: string;
        imageUrl?: string;
    };
    timestamp: number;
}


// interface ChatMessage {
//     role: 'user' | 'assistant';
//     content: string
// }

export interface Model {
    id: string;
    name: string
}